Removed code duplications.
